requirements.txt
Django==5.1.3
djangorestframework==3.15.2
psycopg2-binary==2.9.9
python-dotenv==1.0.*
djangorestframework-simplejwt==5.3.1
factory-boy==3.3.1